+++ djangopypi2/apps/pypi_config/admin.py
@@ -1,7 +1,0 @@
-from django.contrib import admin
-from . import models
-
-admin.site.register(models.GlobalConfiguration)
-admin.site.register(models.MirrorSite)
-admin.site.register(models.MirrorLog)
-