+++ djangopypi2/apps/pypi_frontend/admin.py
@@ -1,8 +1,5 @@
-from django.conf import settings
 from django.contrib import admin
 from . import models
 
-admin.site.register(models.Package)
-admin.site.register(models.Release)
-admin.site.register(models.Distribution)
-admin.site.register(models.Review)
+admin.site.register(models.MirrorSite)
+admin.site.register(models.MirrorLog)