+++ djangopypi2/apps/pypi_frontend/utils.py
@@ -1,11 +1,9 @@
-import logging
-
-
 def debug(func):
     # @debug is handy when debugging distutils requests
     def _wrapped(*args, **kwargs):
         try:
             return func(*args, **kwargs)
         except:
-            logging.exception("@debug")
+            import pdb
+            pdb.pm()
     return _wrapped