+++ djangopypi2/urls.py
@@ -18,5 +18,6 @@ def static_urls():
 
 urlpatterns = patterns('',
     url(r'', include('djangopypi2.apps.pypi_ui.urls')),
+    url(r'', include('djangopypi2.apps.pypi_packages.urls')),
     url(r'', include('djangopypi2.apps.pypi_frontend.urls')),
 ) + static_urls()