+++ djangopypi2/website/settings.py
@@ -119,8 +119,8 @@ INSTALLED_APPS = (
     'django.contrib.staticfiles',
     'django.contrib.admin',
     'djangopypi2.apps.pypi_ui',
-    'djangopypi2.apps.pypi_config',
     'djangopypi2.apps.pypi_metadata',
+    'djangopypi2.apps.pypi_packages',
     'djangopypi2.apps.pypi_frontend',
 )